Crate musli_common

Common utilities shared by Müsli encodings.

The Reader and Writer traits are defined in here which determined the types that can be used in collaboration with Müsli.

Please refer to https://docs.rs/musli for documentation.

Modules

buffered_writer

A writer which buffers the writes before it outputs it into the backing storage.

error

Generic error types that can be used for most Reader / Writer implementations.

fixed_bytes

A container which can store up to a fixed number of uninitialized bytes on the stack and read into and from it.

int

Traits and utilities for dealing with integers.

reader

Trait for governing how a particular source of bytes is read.

wrap

Helpers for integrating musli with I/O types like std::io and std::io::Write.

writer

Trait for governing how a particular sink of bytes is written to.