Trait musli_common::int::IntegerEncoding

pub trait IntegerEncoding: Clone + Copy + Debug + Eq + Hash + Ord + PartialEq + PartialOrd + Sealed {
    // Required methods
    fn encode_unsigned<W, T>(writer: W, value: T) -> Result<(), W::Error>
       where W: Writer,
             T: ByteOrderIo;
    fn decode_unsigned<'de, R, T>(reader: R) -> Result<T, R::Error>
       where R: Reader<'de>,
             T: ByteOrderIo;
    fn encode_signed<W, T>(writer: W, value: T) -> Result<(), W::Error>
       where W: Writer,
             T: Signed,
             T::Unsigned: ByteOrderIo;
    fn decode_signed<'de, R, T>(reader: R) -> Result<T, R::Error>
       where R: Reader<'de>,
             T: Signed,
             T::Unsigned: ByteOrderIo<Signed = T>;
}

Trait which governs how integers are encoded in a binary format.

The two common implementations of this is Variable and Fixed.

Required Methods

fn encode_unsigned<W, T>(writer: W, value: T) -> Result<(), W::Error>where W: Writer, T: ByteOrderIo,

Governs how unsigned integers are encoded into a Writer.

fn decode_unsigned<'de, R, T>(reader: R) -> Result<T, R::Error>where R: Reader<'de>, T: ByteOrderIo,

Governs how unsigned integers are decoded from a Reader.

fn encode_signed<W, T>(writer: W, value: T) -> Result<(), W::Error>where W: Writer, T: Signed, T::Unsigned: ByteOrderIo,

Governs how signed integers are encoded into a Writer.

fn decode_signed<'de, R, T>(reader: R) -> Result<T, R::Error>where R: Reader<'de>, T: Signed, T::Unsigned: ByteOrderIo<Signed = T>,

Governs how signed integers are decoded from a Reader.