Trait musli_common::reader::PosReader

pub trait PosReader<'de>: Reader<'de> {
    type PosMut<'this>: PosReader<'de, Error = Self::Error>
       where Self: 'this;

    // Required methods
    fn pos(&self) -> usize;
    fn pos_borrow_mut(&mut self) -> Self::PosMut<'_>;
}

A reader where the current position is exactly known.

Required Associated Types

type PosMut<'this>: PosReader<'de, Error = Self::Error> where Self: 'this

Reborrowed type.

Why oh why would we want to do this over having a simple &'this mut T?

We want to avoid recursive types, which will blow up the compiler. And the above is a typical example of when that can go wrong. This ensures that each call to borrow_mut dereferences the Reader at each step to avoid constructing a large muted type, like &mut &mut &mut SliceReader<'de>.

Required Methods

fn pos(&self) -> usize

The exact position of a reader.

fn pos_borrow_mut(&mut self) -> Self::PosMut<'_>

Reborrow the current reader.

Object Safety

This trait is not object safe.