Module musli_common::wrap

Helpers for integrating musli with I/O types like std::io and std::io::Write.

The central function in this module is the wrap function which constructs an adapter around an I/O type to work with musli.

Structs

Wrap

Wrapper constructed with wrap.

Functions

wrap

Wrap a type so that it implements Reader or Writer as appropriate.