Trait musli_common::writer::Writer

pub trait Writer {
    type Error: Error;
    type Mut<'this>: Writer<Error = Self::Error>
       where Self: 'this;

    // Required methods
    fn borrow_mut(&mut self) -> Self::Mut<'_>;
    fn write_bytes(&mut self, bytes: &[u8]) -> Result<(), Self::Error>;

    // Provided methods
    fn write_byte(&mut self, b: u8) -> Result<(), Self::Error> { ... }
    fn write_array<const N: usize>(
        &mut self,
        array: [u8; N]
    ) -> Result<(), Self::Error> { ... }
}

The trait governing how a writer works.

Required Associated Types

type Error: Error

The error type raised by the writer.

type Mut<'this>: Writer<Error = Self::Error> where Self: 'this

Reborrowed type.

Why oh why would we want to do this over having a simple &'this mut T?

We want to avoid recursive types, which will blow up the compiler. And the above is a typical example of when that can go wrong. This ensures that each call to borrow_mut dereferences the Reader at each step to avoid constructing a large muted type, like &mut &mut &mut VecWriter.

Required Methods

fn borrow_mut(&mut self) -> Self::Mut<'_>

Reborrow the current type.

fn write_bytes(&mut self, bytes: &[u8]) -> Result<(), Self::Error>

Write bytes to the current writer.

Provided Methods

fn write_byte(&mut self, b: u8) -> Result<(), Self::Error>

Write a single byte.

fn write_array<const N: usize>( &mut self, array: [u8; N] ) -> Result<(), Self::Error>

Write an array to the current writer.

Object Safety

This trait is not object safe.