Module musli_common::int

Traits and utilities for dealing with integers.

In particular the Signed and Unsigned traits are defined in here which allows for musli to work over signed and unsigned numbers generically.

We also have a 7-bit continuation encoding, and zigzag encoding which are based on these.

Modules

continuation

A variable-length 7-bit encoder where each bit indicates if there is a continuation of the sequence or not.

zigzag

Generic zigzag encoding for integers.

Structs

Fixed

A fixed-length integer encoding which encodes something to a little-endian encoding.

FixedUsize

A fixed-length encoding which encodes numbers to the width of L and the endianness of B.

Enums

BigEndian

Defines big-endian serialization.

LittleEndian

Defines little-endian serialization.

Variable

Type that indicates that the given numerical type should use variable-length encoding.

Traits

ByteOrder

Trait governing byte orders.

ByteOrderIo

Helper trait for performing I/O over Unsigned types.

IntegerEncoding

Trait which governs how integers are encoded in a binary format.

Signed

Trait that encodes common behaviors of signed numbers.

Unsigned

Trait that encodes common behaviors of unsigned numbers.

UsizeEncoding

Encoding formats which ensure that variably sized types (like usize, isize) are encoded in a format which is platform-neutral.

Type Aliases

NativeEndian

Defines system native-endian byte order.

NetworkEndian

Defines the network byte order, which is the same as BigEndian.